service 만들ì§기전 권한처리

+++ src/main/java/shop/mtcoding/blog/board/BoardController.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import shop.mtcoding.blog._core.errors.exception.Exception403;
 import shop.mtcoding.blog.user.User;
 
 import java.util.List;
@@ -22,7 +23,11 @@ public class BoardController {
     @Transactional
     @PostMapping("/board/{id}/update")
     public String update(@PathVariable Integer id,BoardRequest.UpdateDTO reqDTO){
-
+        User user = (User) session.getAttribute("sesionUser");
+        Board board = boardRepository.fintById(id);
+        if (user.getId() != board.getUser().getId() ){
+            throw new Exception403("게시글의 권한이 없습니다");
+        }
         boardRepository.updateById(id,reqDTO);
         return "redirect:/board/"+id;
     }
@@ -36,6 +41,12 @@ public class BoardController {
 
     @PostMapping("/board/{id}/delete")
     public String delete(@PathVariable Integer id){
+        User user = (User)session.getAttribute("sessionUser");
+        Board board = boardRepository.fintById(id);
+
+        if (user.getId() != board.getUser().getId()){
+            throw new Exception403("삭제할 권한이 없습니다.");
+        }
         boardPersistRepository.deleteById(id);
 
         return "redirect:/";
@@ -66,6 +77,15 @@ public class BoardController {
     @GetMapping("/board/{id}")
     public String detail(@PathVariable Integer id, HttpServletRequest request) {
         Board board = boardRepository.fintById(id);
+        User user = (User) session.getAttribute("sessionUser");
+        boolean isOwner = false;
+        if (user != null){
+            if (user.getId() == board.getUser().getId()){
+                isOwner = true;
+            }
+        }
+
+        request.setAttribute("isOwner",isOwner);
         request.setAttribute("board",board);
         return "board/detail";
     }
